Convert all flash store reads and write to 32-bit operations
Convert all flash store reads and write to 32-bit operations
They were 8 or 16 bit operation.

+++ TFT/src/User/Hal/gd32f20x/HAL_Flash.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-void HAL_FlashRead(uint8_t * data, uint32_t len);
-void HAL_FlashWrite(uint8_t * data, uint32_t len);
+void HAL_FlashRead(uint32_t * data, uint32_t len);
+void HAL_FlashWrite(uint32_t * data, uint32_t len);
 
 #endif
